Simulate a dataset compliant with AMS-I.F

Generates a tidy dataset representing electricity supplied by renewable generation assets to captive and mini-grid consumers.

simulate_ams_if_dataset(n_grids = 10, n_periods = 12, start_year = 2023, start_month = 1, mean_supply_mwh = 18000, sd_supply_mwh = 2500, baseline_emission_factor = 0.7, project_emission_factor = 0.02, captive_use_share_mean = 0.7, captive_use_share_sd = 0.08)

Arguments

n_grids

Number of captive systems or mini-grids to simulate.

n_periods

Number of monitoring periods per mini-grid.

start_year

Calendar year for the first monitoring period.

start_month

Calendar month (1-12) for the first monitoring period.

mean_supply_mwh

Mean annual electricity supplied per mini-grid in MWh.

sd_supply_mwh

Standard deviation of annual electricity supplied in MWh.

baseline_emission_factor

Baseline emission factor in tCO2e/MWh.

project_emission_factor

Project emission factor in tCO2e/MWh.

captive_use_share_mean

Mean share of electricity consumed within the captive system.

captive_use_share_sd

Standard deviation of the captive use share.

Value

A tibble containing grid identifiers, monitoring metadata, electricity supply, emissions, and captive use shares.

Examples

simulate_ams_if_dataset(n_grids = 5)
#> # A tibble: 60 × 16
#>    grid_id monitoring_period  year month   day monitoring_date monitoring_label
#>    <chr>               <int> <dbl> <dbl> <int> <date>          <chr>           
#>  1 grid_1                  1  2023     1     4 2023-01-04      2023-01         
#>  2 grid_1                  2  2023     2    16 2023-02-16      2023-02         
#>  3 grid_1                  3  2023     3     2 2023-03-02      2023-03         
#>  4 grid_1                  4  2023     4     3 2023-04-03      2023-04         
#>  5 grid_1                  5  2023     5    25 2023-05-25      2023-05         
#>  6 grid_1                  6  2023     6    22 2023-06-22      2023-06         
#>  7 grid_1                  7  2023     7    11 2023-07-11      2023-07         
#>  8 grid_1                  8  2023     8     7 2023-08-07      2023-08         
#>  9 grid_1                  9  2023     9    21 2023-09-21      2023-09         
#> 10 grid_1                 10  2023    10     6 2023-10-06      2023-10         
#> # ℹ 50 more rows
#> # ℹ 9 more variables: electricity_mwh <dbl>, baseline_emission_factor <dbl>,
#> #   project_emission_factor <dbl>, baseline_electricity_mwh <dbl>,
#> #   project_electricity_mwh <dbl>, baseline_emissions_tco2e <dbl>,
#> #   project_emissions_tco2e <dbl>, emission_reductions_tco2e <dbl>,
#> #   captive_use_share <dbl>